binupdate improvements (#648)
* binupdate improvements

- new safeDownload() in helperFunctions
- introduced showProgress argument for emulator downloads
- modified binupdate.sh to use showProgress
- modified emulator/tools install functions to use safeDownload()/showProgress and return proper exit code (CemuNative, Cemu, mGBA, PCSX2-QT, Ryujinx, Vita3K, Xenia, Yuzu, ESDE, SRM)
- installEmuAI() and installEmuBI() are now using new safeDownload() and return proper exit code
- installEmuAI() and installEmuBI() can now write last version info (new arguments)
- yuzu launcher: Added ~/emudeck/yuzu.noupdate to skip yuzu update, zenity progress bar and zenity error dialog

* formatting/whitespaces only

* Fixed shell error (quotes)

---------

+++ functions/EmuScripts/emuDeckVita3K.sh
@@ -14,9 +14,13 @@ Vita3K_cleanup(){
 #Install
 Vita3K_install(){
     echo "Begin Vita3K Install"
-    installEmuBI "Vita3K"  "$(getReleaseURLGH "Vita3K/Vita3K" "ubuntu-latest.zip")" "Vita3K" "zip"
-    unzip -o "$HOME/Applications/Vita3K.zip" -d "$Vita3K_emuPath" && rm -rf "$HOME/Applications/Vita3K.zip"
-    chmod +x "$Vita3K_emuPath/Vita3K"
+    local showProgress="$1"
+    if installEmuBI "Vita3K" "$(getReleaseURLGH "Vita3K/Vita3K" "ubuntu-latest.zip")" "Vita3K" "zip" "$showProgress"; then
+        unzip -o "$HOME/Applications/Vita3K.zip" -d "$Vita3K_emuPath" && rm -rf "$HOME/Applications/Vita3K.zip"
+        chmod +x "$Vita3K_emuPath/Vita3K"
+    else
+        return 1
+    fi
 }
 
 #ApplyInitialSettings
@@ -24,20 +28,18 @@ Vita3K_init(){
     echo "Begin Vita3K Init"
 
     configEmuAI "Vita3K" "config" "$HOME/Applications/Vita3K" "$EMUDECKGIT/configs/Vita3K" "true"
-    
     Vita3K_setEmulationFolder
     Vita3K_setupStorage
     Vita3K_setupSaves #?
     Vita3K_finalize
-
 }
 
 #update
 Vita3K_update(){
     echo "Begin Vita3K update"
-    
+
     configEmuAI "yuzu" "config" "$HOME/.config/Vita3K" "$EMUDECKGIT/configs/Vita3K"
-    
+
     Vita3K_setEmulationFolder
     Vita3K_setupStorage
     Vita3K_setupSaves #?
@@ -51,14 +53,13 @@ Vita3K_setEmulationFolder(){
     echo "Begin Vita3K Path Config"
 
     local prefpath_directoryOpt='pref-path: '
-	local newprefpath_directoryOpt="$prefpath_directoryOpt""$storagePath/Vita3K/"
-	changeLine "$prefpath_directoryOpt" "$newprefpath_directoryOpt" "$Vita3K_configFile"
-
+    local newprefpath_directoryOpt="$prefpath_directoryOpt""$storagePath/Vita3K/"
+    changeLine "$prefpath_directoryOpt" "$newprefpath_directoryOpt" "$Vita3K_configFile"
 }
 
 #SetupSaves
 Vita3K_setupSaves(){
-    echo "Begin Vita3K save link" 
+    echo "Begin Vita3K save link"
     #moveSaveFolder Vita3K saves ??????
 }
 
@@ -124,13 +125,13 @@ Vita3K_finalize(){
 }
 
 Vita3K_IsInstalled(){
-	if [ -e "$Vita3K_emuPath/Vita3K" ]; then
-		echo "true"
-	else
-		echo "false"
-	fi
+    if [ -e "$Vita3K_emuPath/Vita3K" ]; then
+        echo "true"
+    else
+        echo "false"
+    fi
 }
 
 Vita3K_resetConfig(){
-	Vita3K_init &>/dev/null && echo "true" || echo "false"
+    Vita3K_init &>/dev/null && echo "true" || echo "false"
 }
